# laravel-json-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TekManager Sistema de Gestão Integrado",
        "description": "Documentação geral da API de acesso aos dados",
        "contact": {
            "email": "cesar@berci.com.br"
        },
        "license": {
            "name": "TekSea ₢ 2024",
            "url": "http://www.teksea.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v2/estoques": {
            "get": {
                "summary": "Get all estoques",
                "description": "Tabela que armazena informações sobre os itens no estoque",
                "operationId": "bab36cd6b39a748444c62adcc54b393c",
                "responses": {
                    "200": {
                        "description": "Get all estoques"
                    }
                }
            },
            "post": {
                "summary": "Create a new estoque",
                "operationId": "c0c982b1c4c723f90dde900747ef13e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Identificador único de cada item no estoque",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ref_sistema": {
                                        "description": "Referência do sistema para o item",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "ncm": {
                                        "description": "Código NCM (Nomenclatura Comum do Mercosul) do item",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "produto": {
                                        "description": "Descrição detalhada do produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "created_at": {
                                        "description": "Data de criação do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    },
                                    "updated_at": {
                                        "description": "Data de atualização do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new estoque"
                    }
                }
            }
        },
        "/api/v2/estoques/{id}": {
            "get": {
                "summary": "Get a estoque by ID",
                "operationId": "6cc6366fb1750ca1a2163e6c692e4a8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a estoque by ID"
                    }
                }
            },
            "put": {
                "summary": "Update a estoque",
                "operationId": "5d69a5882fce79872147923bc035b546",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Identificador único de cada item no estoque",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ref_sistema": {
                                        "description": "Referência do sistema para o item",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "ncm": {
                                        "description": "Código NCM (Nomenclatura Comum do Mercosul) do item",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "produto": {
                                        "description": "Descrição detalhada do produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "created_at": {
                                        "description": "Data de criação do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    },
                                    "updated_at": {
                                        "description": "Data de atualização do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a estoque"
                    }
                }
            },
            "delete": {
                "summary": "Delete a estoque",
                "operationId": "eadd57d9170053395d6b58c6445e2583",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a estoque"
                    }
                }
            }
        },
...
    }
}